pyproject.toml
[project]
name = "ml-billing"
version = "0.1.0"
description = ""
authors = [{ name = "Dmitriy", email = "alshevskiy.dima@mail.ru" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
packages = [{ include = "billing" }]
optional-dependencies = { dev = [
    "black >=25.1.0,<26.0.0",
    "isort >=6.0.1,<7.0.0",
    "flake8 >=7.2.0,<8.0.0",
    "mypy >=1.15.0,<2.0.0",
    "pylint >=3.3.7,<4.0.0",
    "pytest >=8.3.5,<9.0.0",
    "autoflake >=2.3.1"
]}
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "sqlalchemy (>=2.0.41,<3.0.0)", "pyjwt (>=2.10.1,<3.0.0)", "passlib (>=1.7.4,<2.0.0)", "autopep8 (>=2.3.2,<3.0.0)", "uvicorn (>=0.34.2,<0.35.0)", "celery (>=5.5.2,<6.0.0)", "python-jose (>=3.4.0,<4.0.0)", "langchain-gigachat (>=0.3.10,<0.4.0)", "streamlit (>=1.45.1,<2.0.0)", "st-cookies-manager (>=0.2.2,<0.3.0)", "streamlit-cookies-controller (>=0.0.4,<0.0.5)", "pydantic[email] (>=2.11.4,<3.0.0)", "flower (>=2.0.1,<3.0.0)", "redis (>=6.1.0,<7.0.0)", "docker (>=7.1.0,<8.0.0)", "bcrypt (>=4.3.0,<5.0.0)", "watchdog (>=6.0.0,<7.0.0)", "pycryptodome (>=3.23.0,<4.0.0)", "python-dotenv (>=1.1.0,<2.0.0)", "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
